team-22/Samarthya/requirements.txt
asgiref==3.4.1
beautifulsoup4==4.9.3
bs4==0.0.1
certifi==2021.5.30
cffi==1.14.6
chardet==4.0.0
click==8.0.1
cryptography==3.4.7
Deprecated==1.2.12
Django==3.2.5
englisttohindi==4.1.0
gcloud==0.18.3
googleapis-common-protos==1.53.0
gTTS==2.2.3
httplib2==0.19.1
idna==2.10
importlib-metadata==4.6.1
jwcrypto==0.9.1
keyboard==0.13.5
numpy==1.19.5
oauth2client==4.1.3
pandas==1.1.5
Pillow==8.3.1
playsound==1.2.2
protobuf==3.17.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pycryptodome==3.10.1
pyparsing==2.4.7
PyPDF2==1.26.0
Pyrebase4==4.5.0
pytesseract==0.3.8
python-dateutil==2.8.1
python-jwt==3.3.0
pyttsx3==2.90
pytz==2021.1
requests==2.25.1
requests-toolbelt==0.9.1
rsa==4.7.2
six==1.16.0
soupsieve==2.2.1
sqlparse==0.4.1
typing-extensions==3.10.0.0
urllib3==1.26.6
wrapt==1.12.1
zipp==3.5.0
